Compress archive base class
