+++ src/Files.App/Actions/Content/Archives/Compress/CompressIntoSevenZipAction.cs
@@ -1,0 +1,33 @@
+﻿// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+namespace Files.App.Actions
+{
+	internal sealed class CompressIntoSevenZipAction : BaseCompressArchiveAction
+	{
+		public override string Label
+			=> string.Format("CreateNamedArchive".GetLocalizedResource(), $"{ArchiveHelpers.DetermineArchiveNameFromSelection(context.SelectedItems)}.7z");
+
+		public override string Description
+			=> "CompressIntoSevenZipDescription".GetLocalizedResource();
+
+		public CompressIntoSevenZipAction()
+		{
+		}
+
+		public override Task ExecuteAsync()
+		{
+			var (sources, directory, fileName) = ArchiveHelpers.GetCompressDestination(context.ShellPage);
+
+			IArchiveCreator creator = new ArchiveCreator
+			{
+				Sources = sources,
+				Directory = directory,
+				FileName = fileName,
+				FileFormat = ArchiveFormats.SevenZip,
+			};
+
+			return ArchiveHelpers.CompressArchiveAsync(creator);
+		}
+	}
+}
