--- src/Files.App/Actions/Content/Archives/CompressIntoZipAction.cs
@@ -1,63 +1,0 @@
-﻿// Copyright (c) 2023 Files Community
-// Licensed under the MIT License. See the LICENSE.
-
-using Files.App.Utils.Archives;
-
-namespace Files.App.Actions
-{
-	internal class CompressIntoZipAction : ObservableObject, IAction
-	{
-		private readonly IContentPageContext context;
-
-		public string Label
-			=> string.Format("CreateNamedArchive".GetLocalizedResource(), $"{ArchiveHelpers.DetermineArchiveNameFromSelection(context.SelectedItems)}.zip");
-
-		public string Description
-			=> "CompressIntoZipDescription".GetLocalizedResource();
-
-		public bool IsExecutable =>
-			IsContextPageTypeAdaptedToCommand() &&
-			ArchiveHelpers.CanCompress(context.SelectedItems);
-
-		public CompressIntoZipAction()
-		{
-			context = Ioc.Default.GetRequiredService<IContentPageContext>();
-
-			context.PropertyChanged += Context_PropertyChanged;
-		}
-
-		public Task ExecuteAsync()
-		{
-			var (sources, directory, fileName) = ArchiveHelpers.GetCompressDestination(context.ShellPage);
-
-			IArchiveCreator creator = new ArchiveCreator
-			{
-				Sources = sources,
-				Directory = directory,
-				FileName = fileName,
-				FileFormat = ArchiveFormats.Zip,
-			};
-
-			return ArchiveHelpers.CompressArchiveAsync(creator);
-		}
-
-		private bool IsContextPageTypeAdaptedToCommand()
-		{
-			return
-				context.PageType != ContentPageTypes.RecycleBin &&
-				context.PageType != ContentPageTypes.ZipFolder &&
-				context.PageType != ContentPageTypes.None;
-		}
-
-		private void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-		{
-			switch (e.PropertyName)
-			{
-				case nameof(IContentPageContext.SelectedItems):
-					if (IsContextPageTypeAdaptedToCommand())
-						OnPropertyChanged(nameof(IsExecutable));
-					break;
-			}
-		}
-	}
-}
