+++ src/Files.App/Actions/Content/Archives/Decompress/BaseDecompressArchiveAction.cs
