+++ src/Files.App/Actions/Content/Archives/Decompress/DecompressArchive.cs
